pyproject.toml
[tool.poetry]
name = "product-optimizer"
version = "0.1.0"
description = ""
authors = ["Thorne Wolfenbarger <thornecanyon@gmail.com>"]
readme = "README.md"
packages = [{include = "product_optimizer"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
python-dotenv = "^1.0.0"
openai = "^1.3.7"
pandas = "^2.1.3"
sqlalchemy = "^2.0.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
